Fix DateIndexer returning wrong section if there were no entries for the current day
If there were no log entries for the current day, or for days between the current day and the first day with a log entry, DateIndexer would return the wrong section leading to an ArrayIndexOutOfBoundsException.

This version actually checks the date of the section its looking for against each section to find the proper one.