Add Binary updater activity
There's also a temporary way to get to it. It can currently be accessed by taping menu while on the main screen. This will go away once the preferences activity is added back in. Also, if the requester detects a problem with the database, it will turn off database access until the binary is updated.